+++ examples/ug/foobar.sv
@@ -1,0 +1,11 @@
+module foobar(
+    input logic clk
+);
+
+bar ux (
+  .a(a),
+  .b(b),
+  .c(c)
+);
+
+endmodule