+++ examples/ug/foofoobar.sv
@@ -1,0 +1,19 @@
+module foofoobar();
+
+    foobar u2 (
+        .clk(clk)
+    );
+
+    bar ux (
+    .a(a),
+    .b(b),
+    .c(c)
+    );
+
+    foo ux (
+    .a(a),
+    .b(b),
+    .c(c)
+    );
+
+endmodule